fixes too wide name matching
when a perfect match couldn't be found, sicksubs would check if the season and ep
number were available in the download path. It would first check if the season
was there ('2' in name) then if the ep number was present in the substring
after the place where the season was found ('3' in name[name.find('2')]). This
is wrong first of all because it's an off by one error, and second because in
the case of S02E02, the 2 for the episode can also be found in x264 or other
pieces of text. Now, partial season/ep matching is removed completely for safety